+++ examples/basic_async/basic_async_codegen/src/lib.rs
@@ -1,6 +1,10 @@
 // This file was generated with `cornucopia`. Do not modify.
 
 pub mod client;
+#[allow(clippy::all, clippy::pedantic)]
+#[allow(unused_variables)]
+#[allow(unused_imports)]
+#[allow(dead_code)]
 pub mod queries;
 #[allow(clippy::all, clippy::pedantic)]
 #[allow(unused_variables)]
